Extract mechanics review replay loading
